feat: Add overarching app
The app-of-app structure in this repo is very simplified, hardcoding
most values one would like to template in most cases. The current
structure only passes down the targetRevision to simplify development of
this demo.

The app-tree can be deploy by apply example/app.yaml and setting the
desired source revision