feat: add second stream-team "team-b"
This commit adds a second stream team, "team-b" which emulates how a
team-a would use a development cluster.
The project "team-b" includes an additional role "read-write" which
allows updating apps. This way, team members can point their root app to
different branches or manually adjust inputs for testing. Similarly it
does not deploy an application that invokes the entry point in the team
repo.

In the dev cluster, the team is expected to deploy their root app
manually, either through CI or manual interaction with the cluster. The
cluster would either be created by the team (such that content of
_clusterInfo was known beforehand) or CI (in which case the correct
_clusterInfo would be provided).

For the demo, a fitting root app has been provided in example/app.yaml
in the team-b folder.